retrun typ of Method amountFor() changed to double

+++ src/Customer.java
@@ -50,8 +50,8 @@ class Customer {
 		return result;
 	}
 
-	private int amountFor(Rental each) {
-		int thisAmount = 0;
+	private double amountFor(Rental each) {
+		double thisAmount = 0;
 		// determine amounts for each line
 		switch (each.getMovie().getPriceCode()) {
 		case Movie.REGULAR:
